pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vmware-to-proxmox-migration"
version = "0.1.0"
description = "Production-grade VMware ESXi to Proxmox VE migration tool for Linux guests"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "click==8.1.7",
  "Django==5.1.7",
  "PyYAML==6.0.2",
  "gunicorn==22.0.0",
  "whitenoise==6.7.0",
  "paramiko==3.5.0",
  "pyvmomi==8.0.3.0.1",
]

[project.scripts]
vmware-to-proxmox = "vmware_to_proxmox.cli:cli"
